+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.5.0] - 2019-04-28
+
+### Changed
+
+- Pass blockId to ListContent query.
+
+### Removed
+
+- Remove layout props editions.
+- Disable reorder components.
+
+### Fixed
+
+- Correctly reset content when discarding.
+
 ## [3.4.2] - 2019-04-26
 
 ### Changed